Make an abstract class for CustomCommands
