+++ commands/toolbox_tools/stap 5 - Resultaten nabewerken/calc_structure_statistics.py
@@ -8,9 +8,10 @@ import os
 from ThreeDiToolbox.stats.ncstats import NcStats
 from ThreeDiToolbox.utils.user_messages import pop_up_info
 from ThreeDiToolbox.views.tool_dialog import ToolDialogWidget
+from ThreeDiToolbox.commands.base.custom_command import CustomCommandBase
 
 
-class CustomCommand(object):
+class CustomCommand(CustomCommandBase):
 
     class Fields(object):
         name = "Test script"
@@ -34,25 +35,8 @@ class CustomCommand(object):
         self.layer = None
         self.datasource = None
 
-    def load_defaults(self):
-        """If you only want to use run_it without show_gui, you can try calling
-        this method first to set some defaults.
-
-        This method will try to load the first datasource and the current QGIS
-        layer.
-        """
-        try:
-            self.datasource = self.ts_datasource.rows[0]
-        except IndexError:
-            pop_up_info("No datasource found. Aborting.", title='Error')
-            return
-
-        # Current layer information
-        self.layer = self.iface.mapCanvas().currentLayer()
-        if not self.layer:
-            pop_up_info("No layer selected, things will not go well..",
-                        title='Error')
-            return
+    def run(self):
+        self.show_gui()
 
     def show_gui(self):
         self.tool_dialog_widget = ToolDialogWidget(
@@ -73,9 +57,8 @@ class CustomCommand(object):
         layer_name = self.layer.name()
         structures = ['weir', 'pumpstation', 'pipe', 'orifice', 'culvert']
         if not any(s in layer_name for s in structures):
-            pop_up_info("%s is not a valid structure layer. Valid are: %s" %
-                        (layer_name, structures),
-                        title='Error')
+            pop_up_info("%s is not a valid structure layer. Valid layers are: "
+                        "%s" % (layer_name, structures), title='Error')
             return
 
         result_dir = os.path.dirname(self.datasource.file_path.value)
@@ -106,4 +89,4 @@ class CustomCommand(object):
             for fid, val_dict in result.items():
                 writer.writerow(val_dict)
 
-        pop_up_info("Generated: %s" % filepath)
+        pop_up_info("Generated: %s" % filepath, title='Finished')
